changed Integer for long

+++ src/main/java/com/adoptaamor/adoptaamor/repositories/SavedUserPetsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.adoptaamor.adoptaamor.models.SavedUserPets;
 
-public interface  SavedUserPetsRepository extends JpaRepository <SavedUserPets, Integer> {
+public interface  SavedUserPetsRepository extends JpaRepository <SavedUserPets, Long> {
     
 }
